refactor: Simplify Border by moving shared logic to shared file
