+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.Android.cs
@@ -19,13 +19,6 @@ namespace Microsoft.UI.Xaml.Controls;
 
 public partial class Border
 {
-	private readonly BorderLayerRenderer _borderRenderer;
-
-	public Border()
-	{
-		_borderRenderer = new BorderLayerRenderer(this);
-	}
-
 	partial void OnChildChangedPartial(UIElement previousValue, UIElement newValue)
 	{
 		if (previousValue != null)
@@ -40,20 +33,6 @@ public partial class Border
 	}
 
 	protected override void OnDraw(Android.Graphics.Canvas canvas) => AdjustCornerRadius(canvas, CornerRadius);
-
-	private void UpdateBorder() => _borderRenderer.Update();
-
-	partial void OnBorderBrushChangedPartial() => UpdateBorder();
-
-	protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs e) => UpdateBorder();
-
-	partial void OnBackgroundSizingChangedPartial(DependencyPropertyChangedEventArgs e) => UpdateBorder();
-
-	partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue) => UpdateBorder();
-
-	partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue) => UpdateBorder();
-
-	partial void OnCornerRadiusUpdatedPartial(CornerRadius oldValue, CornerRadius newValue) => UpdateBorder();
 
 	bool ICustomClippingElement.AllowClippingToLayoutSlot => !(Child is UIElement ue) || ue.RenderTransform == null;
 	bool ICustomClippingElement.ForceClippingToLayoutSlot => CornerRadius != CornerRadius.None;
