+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.crossruntime.cs
@@ -20,13 +20,6 @@ namespace Microsoft.UI.Xaml.Controls;
 
 partial class Border
 {
-	private readonly BorderLayerRenderer _borderRenderer;
-
-	public Border()
-	{
-		_borderRenderer = new BorderLayerRenderer(this);
-	}
-
 	partial void OnChildChangedPartial(View previousValue, View newValue)
 	{
 		if (previousValue != null)
@@ -37,24 +30,5 @@ partial class Border
 		AddChild(newValue);
 	}
 
-	private void UpdateBorder() => _borderRenderer.Update();
-
-	partial void OnBorderBrushChangedPartial() => UpdateBorder();
-
-	partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue) =>
-		UpdateBorder();
-
-	partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue) =>
-		UpdateBorder();
-
-	partial void OnCornerRadiusUpdatedPartial(CornerRadius oldValue, CornerRadius newValue) =>
-		UpdateBorder();
-
-	protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs args)
-	{
-		base.OnBackgroundChanged(args);
-
-		UpdateBorder();
-		UpdateHitTest();
-	}
+	partial void OnBackgroundChangedPartial() => UpdateHitTest();
 }
