+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.iOSmacOS.cs
@@ -22,13 +22,6 @@ namespace Microsoft.UI.Xaml.Controls;
 
 public partial class Border
 {
-	private readonly BorderLayerRenderer _borderRenderer;
-
-	public Border()
-	{
-		_borderRenderer = new BorderLayerRenderer(this);
-	}
-
 	partial void OnChildChangedPartial(UIElement previousValue, UIElement newValue)
 	{
 		previousValue?.RemoveFromSuperview();
@@ -47,9 +40,8 @@ public partial class Border
 		UpdateBorder();
 	}
 
-	private void UpdateBorder()
+	partial void AfterUpdateBorderPartial()
 	{
-		_borderRenderer.Update();
 		if (_borderRenderer.BoundsPath is CGPath updated) // UpdateLayer may return null if there is no update
 		{
 			BoundsPath = updated;
@@ -60,16 +52,6 @@ public partial class Border
 
 		this.SetNeedsDisplay();
 	}
-
-	partial void OnBorderBrushChangedPartial() => UpdateBorder();
-
-	partial void OnBorderThicknessChangedPartial(Thickness oldValue, Thickness newValue) => UpdateBorder();
-
-	partial void OnPaddingChangedPartial(Thickness oldValue, Thickness newValue) => UpdateBorder();
-
-	partial void OnCornerRadiusUpdatedPartial(CornerRadius oldValue, CornerRadius newValue) => UpdateBorder();
-
-	protected override void OnBackgroundChanged(DependencyPropertyChangedEventArgs args) => UpdateBorder();
 
 	bool ICustomClippingElement.AllowClippingToLayoutSlot => CornerRadius == CornerRadius.None && (!(Child is UIElement ue) || ue.RenderTransform == null);
 	bool ICustomClippingElement.ForceClippingToLayoutSlot => false;
